ДП2-1 итерация

+++ ewm-main/src/main/java/ru/practicum/ewm/category/dto/CategoryDto.java
@@ -17,8 +17,7 @@ import javax.validation.constraints.Size;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class CategoryDto {
 
-    Integer id;
-
+    Long id;
     @NotBlank
     @Size(min = 1, max = 50)
     String name;
