+++ ewm-main/src/main/java/ru/practicum/ewm/compilation/repository/CompilationRepository.java
@@ -6,7 +6,7 @@ import ru.practicum.ewm.compilation.model.Compilation;
 
 import java.util.List;
 
-public interface CompilationRepository extends JpaRepository<Compilation, Integer> {
-
+public interface CompilationRepository extends JpaRepository<Compilation, Long> {
     List<Compilation> findAllByPinned(Boolean pinned, Pageable pageable);
+
 }
